Have updated all 4 servers to handle STOR and Delete, not tested yet

+++ ds2/ds2.py
@@ -79,14 +79,15 @@ def retr_file(fname, midsock):
 
 def dele_file(fname, midsock):
     try:
-        file_path = 'storage/' + fname
-        if os.path.exists(file_path):
+        file_path = 'C/' + fname
+        file_path1 = 'D/' + fname
+        if os.path.exists(file_path) and os.path.exists(file_path1):
             os.remove(file_path)
+            os.remove(file_path1)
             send_data(fname +' deleted', 'OKOK', midsock)
     except:
         send_data('Error Deleting', 'ERRO', midsock)
     #Fill protocol here
-    
 
 def decode_cmd(cmd, data, midsock):
